fixed logout in new theme
